# src/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Api de Código Postales",
    "description": "Es un api de los códigos postales de SEPOMEX",
    "contact": {
      "name": "Víctor Martínez",
      "url": "mailto:ahal_tocob@hotmail.com"
    },
    "version": "v3"
  },
  "paths": {
    "/api/CodigosPostales/Estados": {
      "get": {
        "tags": ["CodigosPostales"],
        "summary": "Lista de estados",
        "operationId": "Estados",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Estado"
                }
              }
            }
          }
        }
      }
    },
    "/api/CodigosPostales/Estados/{estado}/Alcaldias": {
      "get": {
        "tags": ["CodigosPostales"],
        "summary": "Lista de municipios por estado",
        "parameters": [
          {
            "name": "estado",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Alcaldia"
                }
              }
            }
          }
        }
      }
    },
    "/api/CodigosPostales/Estados/{estado}/Alcaldias/{alcaldia}": {
      "get": {
        "tags": ["CodigosPostales"],
        "summary": "Códigos por estado y alcaldia",
        "parameters": [
          {
            "name": "estado",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "alcaldia",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CodigoPostalDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/CodigosPostales/{codigoPostal}": {
      "get": {
        "tags": ["CodigosPostales"],
        "summary": "Obtener la lista de codigos postales",
        "parameters": [
          {
            "name": "codigoPostal",
            "in": "path",
            "required": true,
            "schema": {
              "maxLength": 5,
              "minLength": 5,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CodigoPostalDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/CodigosPostales/{asentamiento}/Buscar": {
      "get": {
        "tags": ["CodigosPostales"],
        "summary": "Obtener los codigos psotales a partir del nombre de una colonia",
        "parameters": [
          {
            "name": "asentamiento",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CodigoPostalDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/CodigosPostales/Estados/{estado}/Alcaldias/{alcaldia}/{asentamiento}/Buscar": {
      "get": {
        "tags": ["CodigosPostales"],
        "summary": "Obtener los codigos psotales a partir del nombre de una colonia",
        "parameters": [
          {
            "name": "estado",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "alcaldia",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "asentamiento",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CodigoPostalDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/CodigosPostales/Aleatorio": {
      "get": {
        "tags": ["CodigosPostales"],
        "summary": "Obtiene un codigo postal aleatorio",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CodigoPostalDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/CodigosPostales/Estados/{estado}/Aleatorio": {
      "get": {
        "tags": ["CodigosPostales"],
        "summary": "Obtiene un codigo postal aleatorio",
        "parameters": [
          {
            "name": "estado",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CodigoPostalDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Tests/HolaMundo": {
      "get": {
        "tags": ["Tests"],
        "summary": "Hola mundo",
        "parameters": [
          {
            "name": "saludo",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Alcaldia": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "nombre": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CodigoPostalDto": {
        "type": "object",
        "properties": {
          "codigoPostal": {
            "type": "string",
            "nullable": true
          },
          "alcaldiaId": {
            "type": "integer",
            "format": "int32"
          },
          "estado": {
            "type": "string",
            "nullable": true
          },
          "estadoId": {
            "type": "integer",
            "format": "int32"
          },
          "alcaldia": {
            "type": "string",
            "nullable": true
          },
          "tipoDeAsentamiento": {
            "type": "string",
            "nullable": true
          },
          "asentamiento": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Estado": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "nombre": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}